Crate linked_list_allocator

Modules

hole

Structs

Heap

A fixed size heap backed by a linked list of free memory blocks.

LockedHeap

Functions

align_down_size

Align downwards. Returns the greatest x with alignment align so that x <= addr. The alignment must be a power of 2.

align_up

Align upwards. Returns the smallest x with alignment align so that x >= addr. The alignment must be a power of 2.

align_up_size